requirements.txt
fastapi
azure-storage-blob
pandas
jinja2
python-multipart
uvicorn
python-dotenv